Exception handling

+++ back-end/src/main/java/lk/ijse/dep10/pos/business/custom/impl/CustomerBOImpl.java
@@ -43,9 +43,13 @@ public class CustomerBOImpl implements CustomerBO {
         try (Connection connection = dataSource.getConnection()) {
             customerDAO.setConnection(connection);
 
-            if (customerDAO.existsCustomerByContact(customerDTO.getContact())) {
-                throw new BusinessException(BusinessExceptionType.DUPLICATE_RECORD, "Update failed: Contact number: " + customerDTO.getContact() + " already exists");
+            if (customerDAO.existsCustomerByContactAndNotId(customerDTO.getContact(), customerDTO.getId())) {
+                    throw new BusinessException(BusinessExceptionType.DUPLICATE_RECORD, "Update failed: Contact number: " + customerDTO.getContact() + " already exists");
             }
+
+            if (!customerDAO.existsById(customerDTO.getId()))
+                throw new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
+                        "Update failed: Customer ID: " + customerDTO.getId() + " does not exist");
             customerDAO.update(transformer.toCustomerEntity(customerDTO));
         }
     }
@@ -59,6 +63,10 @@ public class CustomerBOImpl implements CustomerBO {
             if (orderCustomerDAO.existsOrderByCustomerId(customerId)) {
                 throw new BusinessException(BusinessExceptionType.INTEGRITY_VIOLATION, "Delete failed: Customer ID: " + customerId + " is already associated with some orders");
             }
+
+            if (!customerDAO.existsById(customerId))
+                throw new BusinessException(BusinessExceptionType.RECORD_NOT_FOUND,
+                        "Delete failed: Customer ID: " + customerId + " does not exist");
 
             customerDAO.deleteById(customerId);
         }
